# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CommonResponse:
    properties:
      data:
        $ref: '#/definitions/dto.Resp'
    type: object
  dto.DataUserResponseSwagger:
    properties:
      data:
        $ref: '#/definitions/dto.UserResponseSwagger'
    type: object
...
  dto.LikeUserSwagger:
    properties:
      isLike:
        type: boolean
    type: object
...
  dto.Resp:
    properties:
      success:
        type: boolean
    type: object
...
  dto.UserResponseSwagger:
    properties:
      avatarUrl:
        type: string
      backgroundUrl:
        type: string
      canSubscribe:
        type: boolean
      collectionItemsCount:
        default: 0
        type: integer
      collectionSubscriptions:
        items:
          type: string
        type: array
      collectionsCount:
        default: 0
        type: integer
      created:
        type: string
      dislikes:
        type: integer
      email:
        type: string
      likes:
        type: integer
      login:
        type: string
      subscribers:
        type: integer
      subscriptions:
        items:
          type: string
        type: array
      telegramId:
        type: string
      totalSum:
        default: 0
        type: integer
      userName:
        type: string
      vkId:
        type: string
    type: object
...
  models.ChangePasswordReset:
    properties:
      password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
...
  models.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.SignInResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  models.SignUpRequest:
    properties:
      bio:
        type: string
      city:
        type: string
      email:
        type: string
      login:
        type: string
      password:
        minLength: 8
        type: string
      profileName:
        type: string
      userName:
        type: string
    required:
    - email
    - login
    - password
    - userName
    type: object
  models.SignUpResponse:
    properties:
      data:
        type: string
    type: object
...
  models.UserRequestUpdate:
    properties:
      avatarUrl:
        type: string
      backgroundUrl:
        type: string
      bio:
        type: string
      city:
        type: string
      email:
        type: string
      profileName:
        type: string
      userName:
        type: string
    type: object
  models.UserRequestUpdateFirstTime:
    properties:
      avatarUrl:
        type: string
      backgroundUrl:
        type: string
      bio:
        type: string
      city:
        type: string
      email:
        type: string
      login:
        type: string
      profileName:
        type: string
      userName:
        type: string
    type: object
...
host: dev.lootor.me
info:
  contact: {}
  description: Документация к API Lootor.me
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lootor
  version: "1.0"
paths:
  /public/auth/reset:
    get:
      consumes:
      - application/json
      parameters:
      - description: email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: сброс пароля
      tags:
      - users
  /public/auth/signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SignInRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SignInResponse'
      summary: Авторизация
      tags:
      - users
  /public/auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SignUpResponse'
      summary: создание пользователя
      tags:
      - users
...
  /public/user:
    get:
      consumes:
      - application/json
      parameters:
      - description: логин
        in: query
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DataUserResponseSwagger'
      summary: получение юзера по логину
      tags:
      - users
  /public/user/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: search
        in: query
        name: search
        required: true
        type: string
      - description: order
        in: query
        name: order
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DataUserResponseSwagger'
      summary: получение юзеров
      tags:
      - users
  /public/user/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordReset'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: смена пароля по токену
      tags:
      - users
...
  /secured/user/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: parolj
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: User data
        in: query
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: Смена пароля
      tags:
      - users
  /secured/user/rating:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LikeUserSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: Лайк/дизлайк
      tags:
      - users
  /secured/user/subscriptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: признак подписки
        in: query
        name: isSubscribe
        required: true
        type: boolean
      - description: User data
        in: query
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: Подписка одного юзера на другого
      tags:
      - users
  /secured/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserRequestUpdate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DataUserResponseSwagger'
      summary: смена некоторых данных пользователя
      tags:
      - users
  /secured/user/update/reg:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserRequestUpdateFirstTime'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SignInResponse'
      summary: смена логина после оауса
      tags:
      - users
...
schemes:
- https
swagger: "2.0"